requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
openpyxl==3.1.2
tkcalendar==1.6.1
